+++ frontend/src/components/EPLExpandedGameInfo/EPLExpandedGameInfo.js
@@ -1,0 +1,7 @@
+import React, { Component } from "react";
+
+export default class EPLExpandedGameInfo extends Component {
+  render() {
+    return <div>EPL Expanded Game Info</div>;
+  }
+}
