+++ frontend/src/components/MLBExpandedGameInfo/MLBExpandedGameInfo.js
@@ -1,0 +1,7 @@
+import React, { Component } from "react";
+
+export default class MLBExpandedGameInfo extends Component {
+  render() {
+    return <div>MLB Expanded Game Info</div>;
+  }
+}
